pip_requirements.txt
numpy
opencv-python
pillow
scipy
matplotlib
